F-36: keep injector text-noise hash modulo in u64 — negative slice crash fixed
hash() % <i64 expr> promoted to i64 and wrapped negative for ~half the hash
range; str.slice rejected the offsets, killing s1 on any corpus with more than
a handful of hallucination victims (this is what made the synthetic baseline
unobtainable). Modulo now runs in u64 (result < span, lossless downcast). The
characterization pin flips from expect-crash to expect-clean-corruption; s1
label/feature pins are unchanged (label assignment used literal-modulo, which
never wrapped).

+++ ars/data/anomalies_injection.py
@@ -78,7 +78,10 @@ class TextNoise:
         ht      = idx.hash(seed = seed + 211)
         hs      = idx.hash(seed = seed + 331)
         hk      = idx.hash(seed = seed)
-        pos     = (hp % span).cast(pl.Int64)
+        # hash() is UInt64; `% <i64 expr>` promotes to i64 and wraps negative for
+        # ~half the hash range, which str.slice then rejects (finding F-36).
+        # Keep the modulo in u64: the result is < span, so the cast is lossless.
+        pos     = (hp % span.cast(pl.UInt64)).cast(pl.Int64)
         sep     = pl.lit(' ')
 
         choose      = lambda opts, h: reduce(
